requirements.txt
seaborn==0.11.1
mmpycocotools==12.0.3
thop==0.0.31.post2005241907
requests==2.25.1
torch==1.7.1
tqdm==4.56.0
onnx==1.8.1
matplotlib==3.3.3
pandas==1.2.0
torchvision==0.8.2
timm==0.4.9
opencv_python==4.5.1.48
numpy==1.19.5
scipy==1.6.0
coremltools==4.1
pafy==0.5.5
Pillow==8.2.0
pycocotools==2.0.2
PyYAML==5.4.1
